requirements.txt
Django~=3.2.8
django-model-utils~=4.1.1
django-modeltranslation~=0.17.5
django-ckeditor~=6.2.0
django-admin-honeypot~=1.1.0
